Starting refactoring material ui into styled components
