+++ src/components/signUpForm/styles/contactInfoStyles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const ContactContainer = styled.div`
+  margin: 24px;
+`
+
+export const FirstName = styled.div``
