+++ src/theme.js
@@ -1,0 +1,12 @@
+export const theme = {
+  palette: {
+    primary: {
+      main: '#517E92',
+      light: '#91BCCF',
+    },
+    secondary: {
+      main: '#FF8D86',
+      light: '#F0B6A7',
+    },
+  },
+}
